DHFPROD-6235: Improving Spark example project
