+++ marklogic-data-hub-spark-connector/spark-test-project/gradle.properties
@@ -1,6 +1,9 @@
-# Connection properties for the test program
+# Connection properties for the test programs
+
 # Can override this via gradle-local.properties, which is gitignored
 host=localhost
-username=test-data-hub-operator
+
+# Assumes a connection to an application deployed from ./examples/reference-entity-model
+username=hub-operator
 password=password
