+++ marklogic-data-hub-spark-connector/spark-test-project/src/main/resources/data/customers.csv
@@ -1,0 +1,6 @@
+CustomerID,Name,Email,SSN
+501,Jane Spark,jane@example.org,501-99-9999
+502,Jim Spark,jim@example.org,502-99-9999
+503,Juanita Spark,juanita@example.org,503-99-9999
+504,Jasmine Spark,jasmine@example.org,504-99-9999
+505,John Spark,john@example.org,505-99-9999
