--- marklogic-data-hub-spark-connector/spark-test-project/src/main/resources/data/databook.csv
@@ -1,10 +1,0 @@
-﻿FirstName,LastName,Details,Cell
-Harry,Potter,Gryffindoor,1234
-Ronald,Wesley,London,5678
-Hermione,Granger,Earth,3456
-Draco,Malfoy,Slytherin,9894
-Albus,Dumbledore,Hogwarts,6765
-Severus,Snape,Slytherine,5566
-Ginny,Waesley,Gryffindoor,7766
-Cedric,Diggory,Goblet of Fire,4433
-Rubeus,Hagrid,Master,9999
