+++ marklogic-data-hub-spark-connector/spark-test-project/src/main/resources/log4j.properties
@@ -4,14 +4,11 @@
 # Root logger option
 log4j.rootLogger=WARN, stdout
 
-log4j.logger.test=INFO, stdout
-log4j.additivity.test=false
+log4j.logger.org.example=INFO, stdout
+log4j.additivity.org.example=false
 
 log4j.logger.com.marklogic.hub=DEBUG, stdout
 log4j.additivity.com.marklogic.hub=false
-
-log4j.logger.com.marklogic.client.dataservices=TRACE, stdout
-log4j.additivity.com.marklogic.client.dataservices=false
 
 # Direct log messages to stdout
 log4j.appender.stdout=org.apache.log4j.ConsoleAppender
